requirements.txt
pandas==2.3.2
scikit-learn==1.7.1
joblib==1.5.2
fastapi==0.116.1
uvicorn[standard]==0.35.0
pyyaml==6.0.2
pytest==8.4.1
python-dotenv==1.1.0
httpx==0.28.1
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
scikit-learn==1.7.1
joblib==1.5.2
googlemaps==4.10.0
numpy==2.4.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
python-dotenv==1.1.0
pyyaml==6.0.2
pytest==8.4.1
httpx==0.28.1
 